+++ src/Application/Import/Import.h
@@ -1,4 +1,5 @@
-#include <QtXml>
+#ifndef IMPORT_H
+#define IMPORT_H
 
 #include "Entity/VocabularyGroup.h"
 
@@ -7,10 +8,13 @@ class Import
 	public:
 		virtual ~Import();
 		static bool importFromFile( QString file );
+		static bool importFromDb( QString dbName );
 
 	protected:
-		static VocabularyGroupPtr createGroup();
+		static VocabularyGroupPtr createGroup( QString groupName = "Imported" );
 
 	private:
 		Import();
 };
+
+#endif	// IMPORT_H